4.21 delete blog

+++ controllers/blogs.js
@@ -3,21 +3,13 @@ const Blog = require('../models/blog.js')
 const User = require('../models/user.js')
 const jwt = require('jsonwebtoken')
 
-blogRouter.get('/', async (request, response) => {
-    console.log(request.headers)
-    console.log(Object.keys(request))
-    
-    
+blogRouter.get('/', async (request, response) => { 
     const blog = await Blog.find({}).populate('userId', {username: 1, name: 1})
     response.json(blog)
 })
 
 blogRouter.post('/', async (request, response) => {
     const body = request.body
-    console.log(request.headers)
-    console.log('------------')
-    
-    console.log(Object.keys(request))    
 
     const decodedToken = jwt.verify(request.token, process.env.SECRET)
     
@@ -47,8 +39,24 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (!decodedToken.id) {
+        return response.status(401).json({error: 'token invalid'})
+    }
+
+    const blog = await Blog.findById(request.params.id)
+
+    if (!blog) {
+        return response.status(401).json({error: 'blog no longer exist'})
+    }
+
+    if (blog.userId.toString() === decodedToken.id) {
+        console.log('same userid')
+        await Blog.findByIdAndDelete(request.params.id)
+        response.status(204).end()
+    } else {response.status(401).json({error: 'invalid username'})}
 })
 
 blogRouter.put('/:id', async (request, response) => {
